requirements.txt
flask-sqlalchemy
pymysql
mysql-connector-python  # Alternative MySQL adapter
flask-login
flask-bcrypt
python-dotenv
